feat: useActiveSection custom hook
